Resolve, strings in code, Add game logic Add player turn display

+++ TwisterSpinner/TwisterSpinnerView.cpp
@@ -48,6 +48,8 @@ CTwisterSpinnerView::CTwisterSpinnerView() noexcept
 
 void CTwisterSpinnerView::Reset()
 {
+	spinButton.EnableWindow(false);
+	removeButton.EnableWindow(false);
 	OnNewGameButtonClicked();
 }
 
@@ -78,46 +80,53 @@ COLORREF giveColor(int i){
 }
 void CTwisterSpinnerView::OnDraw(CDC* pDC)
 {
+	//Init of variables
 
-	pDC->SetBkColor(RGB(200, 200, 200));
 	CRect screen;
 	GetWindowRect(screen);
 	CPoint center = CPoint((screen.right - screen.left) / 2, (screen.bottom - screen.top) / 2);
 	CPoint previous = CPoint(center.x, center.y+r);
-	pDC->MoveTo(previous);
+	pDC->SetBkColor(RGB(200, 200, 200));
 	FlickerFreeDC::CMemDC dc{ pDC };
-	CString stringFetcher[6];
-	for (int i = 0; i < 6; ++i)
+
+	//Drawing labels
+
+	CString stringFetcher[7];
+	for (int i = 0; i < 7; ++i)
 	{
 		stringFetcher[i].LoadString(IDS_PLAYERSTRING+i);
 	}
+	CFont font;
+	int iHeight = -MulDiv(9, ::GetDeviceCaps(dc, LOGPIXELSY), 60);
+	int iWidth = -MulDiv(7, ::GetDeviceCaps(dc, LOGPIXELSX), 90);
+	font.CreateFont(iHeight, iWidth, 0, 0, FW_MEDIUM, FALSE, FALSE, 0, ANSI_CHARSET, OUT_DEFAULT_PRECIS, CLIP_DEFAULT_PRECIS, DEFAULT_QUALITY, DEFAULT_PITCH | FF_MODERN, _T("Comic Sans"));
+	dc.SelectObject(&font);
 	dc.DrawText(stringFetcher[0], CRect(center.x - r - 320, center.y - r, center.x - r - 250, center.y - r + 20), DT_SINGLELINE | DT_LEFT);
 	dc.DrawText(stringFetcher[1], CRect(center.x - r-20, center.y - r, center.x - r + 50, center.x - r + 20), DT_SINGLELINE|DT_LEFT);
 	dc.DrawText(stringFetcher[2], CRect(center.x + r - 20, center.y - r, center.x + r + 50, center.x - r + 20), DT_SINGLELINE | DT_LEFT);
 	dc.DrawText(stringFetcher[3], CRect(center.x - r - 20, center.y + r, center.x - r + 50, center.x + r + 20), DT_SINGLELINE | DT_LEFT);
 	dc.DrawText(stringFetcher[4], CRect(center.x + r - 20, center.y + r, center.x + r + 50, center.x + r + 20), DT_SINGLELINE | DT_LEFT);
 	dc.DrawText(stringFetcher[5], CRect(center.x - r - 370, center.y - r - 60, center.x - r - 250, center.y - r - 40), DT_SINGLELINE | DT_LEFT);
-	CFont font;
+	if (gameStarted) {
+		if (currentIndex == listBox.GetCount()) {
+			currentIndex = 0;
+		}
+		CString playerTurn;
+		listBox.GetText(currentIndex,playerTurn);
+		CString turn;
+		turn.LoadString(IDS_TURN);
+		playerTurn.Append(turn);
+		dc.DrawText(playerTurn, CRect(center.x - 200, center.y - r - 30, center.x + 200, center.x - r - 10), DT_SINGLELINE | DT_CENTER);
+	}
+	CFont fontBig;
+	iHeight = -MulDiv(18, ::GetDeviceCaps(dc, LOGPIXELSY), 72);
+	fontBig.CreateFont(iHeight,0,0,0,FW_SEMIBOLD,FALSE,FALSE,0,ANSI_CHARSET,OUT_DEFAULT_PRECIS,CLIP_DEFAULT_PRECIS,DEFAULT_QUALITY,DEFAULT_PITCH | FF_MODERN,_T("Comic Sans"));
+	dc.SelectObject(&fontBig);
+	dc.DrawText(stringFetcher[6], CRect(center.x - 200, center.y - r - 70, center.x + 200, center.x - r - 50), DT_SINGLELINE | DT_CENTER);
 
-	//TODO:: calculate font, add players turn, remove adding during game, fit all texts
-	font.CreateFont(
-		28,                       // nHeight
-		0,                        // nWidth
-		0,                        // nEscapement
-		0,                        // nOrientation
-		FW_SEMIBOLD,                // nWeight
-		FALSE,                    // bItalic
-		FALSE,                    // bUnderline
-		0,                        // cStrikeOut
-		ANSI_CHARSET,             // nCharSet
-		OUT_DEFAULT_PRECIS,       // nOutPrecision
-		CLIP_DEFAULT_PRECIS,      // nClipPrecision
-		DEFAULT_QUALITY,          // nQuality
-		DEFAULT_PITCH | FF_MODERN, // nPitchAndFamily
-		_T("Comic Sans"));            // lpszFacename
+	//Drawing the wheel
 
-	dc.SelectObject(&font);
-	dc.DrawText(L"Welcome to TwisterSpinner", CRect(center.x - 200, center.y - r - 70, center.x + 200, center.x - r - 50), DT_SINGLELINE | DT_CENTER);
+	dc.MoveTo(previous);
 	CTwisterSpinnerDoc* pDoc = GetDocument();
 	ASSERT_VALID(pDoc);
 	if (!pDoc)
@@ -139,6 +148,9 @@ void CTwisterSpinnerView::OnDraw(CDC* pDC)
 			dc.Pie(CRect(center.x-r,center.y-r,center.x+r,center.y+r), CPoint(x, y),previous);
 			previous=CPoint(x,y);
 		}
+
+		//Draw the arrow
+
 		CPen drawPen(PS_SOLID, 3, RGB(0, 0, 0));
 		double arrowR = r * 0.9;
 		dc.SelectObject(&drawPen);
@@ -175,7 +187,6 @@ void CTwisterSpinnerView::OnSpinButtonClicked()
 		spinAmount -= 2 * M_PI;
 		spinAngle -= 2 * M_PI;
 	}
-	//spinAmount += (1 + (int)(500.0 * rand() / (RAND_MAX + 1.0)))/10+ 2 * pi;
 	std::uniform_real_distribution<double> dist(5*M_PI, 7*M_PI);	
 	spinAmount +=dist(mt);
 	arrowSpun = true;
@@ -194,40 +205,80 @@ void CTwisterSpinnerView::OnAddButtonClicked()
 	}
 	textBox.SetWindowTextW(L"");
 	listBox.AddString(name);
+	removeButton.EnableWindow(true);
 }
 void CTwisterSpinnerView::OnRemoveButtonClicked()
 {
-	listBox.DeleteString(listBox.GetCurSel());
-	if (listBox.GetCount() == 1) {
+	int index = listBox.GetCurSel();
+	if ( index< currentIndex&&gameStarted&&currentIndex!=0) {
+		currentIndex-= 1;
+	}
+	listBox.DeleteString(index);
+	if (listBox.GetCount() == 1&&gameStarted) {
 		CString winner;
 		listBox.GetText(0, winner);
-		winner.Append(L" wins!");
-		MessageBox(winner, L"Congratulations", MB_OK);
+		CString stringFetcher;
+		stringFetcher.LoadString(IDS_WINNER);
+		winner.Append(stringFetcher);
+		stringFetcher.LoadString(IDS_GRATS);
+		MessageBox(winner, stringFetcher, MB_OK);
+		Reset();
 	}
+	Invalidate();
 }
 void CTwisterSpinnerView::OnNewGameButtonClicked()
 {
-	while (listBox.GetCount() > 0) {
-		listBox.DeleteString(0);
+	if (gameStarted)
+	{
+		CString stringFetcher;
+		stringFetcher.LoadString(IDS_GAMEOVER);
+		CString stringFetcher2;
+		stringFetcher2.LoadString(IDS_NEWGAMEQ);
+		if (MessageBox(stringFetcher2,stringFetcher, MB_YESNO) == IDYES)
+		{
+			while (listBox.GetCount() > 0) {
+				listBox.DeleteString(0);
+			}
+			std::uniform_real_distribution<double> dist(0, 2 * M_PI);
+			spinAmount = dist(mt);
+			spinAngle = dist(mt);
+			currentIndex = 0;
+			gameStarted = false;
+			addButton.EnableWindow(true);
+			spinButton.EnableWindow(false);
+			removeButton.EnableWindow(true);
+		}	
 	}
-	std::uniform_real_distribution<double> dist(0, 2 * M_PI);
-	spinAmount = dist(mt);
-	spinAngle = dist(mt);
+	else {
+		if (listBox.GetCount() >= 2) {
+			gameStarted = true;
+			spinButton.EnableWindow(true);
+			addButton.EnableWindow(false);
+			removeButton.EnableWindow(false);
+		}
+	}
 	Invalidate();
+	
 }
 void CTwisterSpinnerView::DisplayLimbAndColor() {
-	CString limbAndColor;
+	CString limbAndColor; 
+	CString stringFetcher;
+	
 	if (cos(spinAngle) < 0) {
-		limbAndColor.Append(L"Left ");
+		stringFetcher.LoadString(IDS_LEFT);
+		limbAndColor.Append(stringFetcher);
 	}
 	else {
-		limbAndColor.Append(L"Right ");
+		stringFetcher.LoadString(IDS_RIGHT);
+		limbAndColor.Append(stringFetcher);
 	}
 	if (sin(spinAngle) <= 0) {
-		limbAndColor.Append(L"hand ");
+		stringFetcher.LoadString(IDS_HAND);
+		limbAndColor.Append(stringFetcher);
 	}
 	else {
-		limbAndColor.Append(L"leg ");
+		stringFetcher.LoadString(IDS_LEG);
+		limbAndColor.Append(stringFetcher);
 	}
 	double angleDeg = (spinAngle*180) / M_PI;
 	while (angleDeg > 90) {
@@ -235,26 +286,32 @@ void CTwisterSpinnerView::DisplayLimbAndColor() {
 	}
 	if (angleDeg >= 3 * 22.5)
 	{
-		limbAndColor.Append(L"yellow");
+		stringFetcher.LoadString(IDS_YELLOW);
+		limbAndColor.Append(stringFetcher);
 	}
 	else if (angleDeg >= 2 * 22.5)
 	{
-		limbAndColor.Append(L"blue");
+		stringFetcher.LoadString(IDS_BLUE);
+		limbAndColor.Append(stringFetcher);
 	}
 	else if (angleDeg >= 22.5)
 	{
-		limbAndColor.Append(L"green");
+		stringFetcher.LoadString(IDS_GREEN);
+		limbAndColor.Append(stringFetcher);
 	}
 	else
 	{
-		limbAndColor.Append(L"red");
+		stringFetcher.LoadString(IDS_RED);
+		limbAndColor.Append(stringFetcher);
 	}
-	MessageBox((LPCTSTR)limbAndColor, L"You spun", MB_OK);
+	stringFetcher.LoadString(IDS_SPUN);
+	MessageBox((LPCTSTR)limbAndColor, stringFetcher, MB_OK);
+	currentIndex += 1;
+	arrowSpun = false;
 	newGameButton.EnableWindow(true);
-	addButton.EnableWindow(true);
 	removeButton.EnableWindow(true);
 	spinButton.EnableWindow(true);
-	arrowSpun = false;
+	Invalidate();
 }
 void CTwisterSpinnerView::OnTimer(UINT_PTR TimerVal)
 {
@@ -302,11 +359,6 @@ CTwisterSpinnerDoc* CTwisterSpinnerView::GetDocument() const // non-debug versio
 	return (CTwisterSpinnerDoc*)m_pDocument;
 }
 #endif //_DEBUG
-
-
-// CTwisterSpinnerView message handlers
-
-
 int CTwisterSpinnerView::OnCreate(LPCREATESTRUCT lpCreateStruct)
 {
 	if (CView::OnCreate(lpCreateStruct) == -1)
@@ -316,11 +368,21 @@ int CTwisterSpinnerView::OnCreate(LPCREATESTRUCT lpCreateStruct)
 	spinAmount=(rand()) / (RAND_MAX / (2 * M_PI));
 	spinAngle = (rand()) / (RAND_MAX / (2 * M_PI));
 
+	CString stringFetcher;
 
-	spinButton.Create(_T("Spin!"), BS_PUSHBUTTON | WS_VISIBLE, CRect(300, 100, 400, 150),this,IDC_SPINBUTTON);
-	addButton.Create(_T("Add player"), BS_PUSHBUTTON | WS_VISIBLE, CRect(300, 100, 400, 150), this, IDC_ADDBUTTON);
-	removeButton.Create(_T("Remove player"), BS_PUSHBUTTON | WS_VISIBLE, CRect(300, 100, 400, 150), this, IDC_REMOVEBUTTON);
-	newGameButton.Create(_T("Start new game"), BS_PUSHBUTTON | WS_VISIBLE, CRect(300, 100, 400, 150), this, IDC_NEWGAMEBUTTON);
+	stringFetcher.LoadString(IDS_SPIN);
+	spinButton.Create(stringFetcher, BS_PUSHBUTTON | WS_VISIBLE, CRect(300, 100, 400, 150),this,IDC_SPINBUTTON);
+	spinButton.EnableWindow(false);
+
+	stringFetcher.LoadString(IDS_ADDPLAYER);
+	addButton.Create(stringFetcher, BS_PUSHBUTTON | WS_VISIBLE, CRect(300, 100, 400, 150), this, IDC_ADDBUTTON);
+	
+	stringFetcher.LoadString(IDS_REMOVEPLAYER);
+	removeButton.Create(stringFetcher, BS_PUSHBUTTON | WS_VISIBLE, CRect(300, 100, 400, 150), this, IDC_REMOVEBUTTON);
+	removeButton.EnableWindow(false);
+	
+	stringFetcher.LoadString(IDS_STARTNEWGAMEB);
+	newGameButton.Create(stringFetcher, BS_PUSHBUTTON | WS_VISIBLE, CRect(300, 100, 400, 150), this, IDC_NEWGAMEBUTTON);
 
 	textBox.Create(WS_CHILD | WS_VISIBLE|WS_BORDER, CRect(100, 100, 200, 500), this, IDC_TEXTBOX);
 	listBox.Create(WS_CHILD | WS_VISIBLE|WS_BORDER,CRect(100,100,200,500),this,IDC_LISTBOX);
@@ -332,16 +394,7 @@ BOOL CTwisterSpinnerView::OnEraseBkgnd(CDC* pDC)
 {
 	return false;
 }
-void CTwisterSpinnerView::OnGetMinMaxInfo(MINMAXINFO* lpMMI)
-{
-		// Adjust the minimum height and width, if needed
-		lpMMI->ptMinTrackSize.x = 1000;
-		lpMMI->ptMinTrackSize.y = 1000;
-		// Adjust the maximum height and width, if necessary
-		lpMMI->ptMaxTrackSize.x = 1000;
-		lpMMI->ptMaxTrackSize.y = 1000;
-		CView::OnGetMinMaxInfo(lpMMI);
-}
+
 
 BOOL CTwisterSpinnerView::OnWndMsg(UINT message, WPARAM wParam, LPARAM lParam, LRESULT* pResult)
 {
